added the sorted_categories.json
sorted_categories.json tells us how many of the events have this category. Scrolling through should help with understanding. This category words are the origin of the colours, they were embedded in to 1000+ dimensional vectors using openAI embeddings, then as a set, PCA reduced to 3D vectors to be able to represent colours that should relate to theme. The amount of times it shows up was not taken in to account for the PCA reduction.